Openshift Assisted Installer
partial-jira-bug: 30720

+++ openshift/functions.sh
@@ -50,31 +50,36 @@ EOF
 }
 
 function download_artefacts() {
-  [[ ! -d ${DOWNLOADS_DIR} ]] && mkdir -p ${DOWNLOADS_DIR}
+  mkdir -p ${DOWNLOADS_DIR}
 
   if [[ ! -f ${DOWNLOADS_DIR}/${CLIENT} ]]; then
+    echo "INFO: download $CLIENT_URL"
     wget -nv "$CLIENT_URL" -O "${DOWNLOADS_DIR}/$CLIENT"
     tar -xf "${DOWNLOADS_DIR}/${CLIENT}"
     rm -f README.md
   fi
-  if [[ ! -f ${DOWNLOADS_DIR}/${INSTALLER} ]]; then
-    wget -nv "$INSTALLER_URL" -O "${DOWNLOADS_DIR}/$INSTALLER"
-    tar -xf "${DOWNLOADS_DIR}/${INSTALLER}"
-    rm -f README.md
+
+  if [[ "${OPENSHIFT_AI_INSTALLER,,}" != 'true' ]]; then
+    if [[ ! -f ${DOWNLOADS_DIR}/${INSTALLER} ]]; then
+      echo "INFO: download $INSTALLER_URL"
+      wget -nv "$INSTALLER_URL" -O "${DOWNLOADS_DIR}/$INSTALLER"
+      tar -xf "${DOWNLOADS_DIR}/${INSTALLER}"
+      rm -f README.md
+    fi
   fi
 
   if [[ "$PROVIDER" == "kvm" ]]; then
-    if [[ ! -f ${DOWNLOADS_DIR}/${RHCOS_IMAGE} ]]; then
-        wget -nv "$RHCOS_URL" -O "${DOWNLOADS_DIR}/${RHCOS_IMAGE}"
-    fi
-    if [[ ! -f ${DOWNLOADS_DIR}/${RHCOS_KERNEL} ]]; then
-        wget -nv "${RHCOS_MIRROR}/${RHCOS_VERSION}/$RHCOS_KERNEL" -O "${DOWNLOADS_DIR}/$RHCOS_KERNEL"
-    fi
-    if [[ ! -f ${DOWNLOADS_DIR}/${RHCOS_INITRAMFS} ]]; then
-        wget -nv "${RHCOS_MIRROR}/${RHCOS_VERSION}/$RHCOS_INITRAMFS" -O "${DOWNLOADS_DIR}/$RHCOS_INITRAMFS"
-    fi
+    local rhcos_base_url="${RHCOS_MIRROR}/${OPENSHIFT_VERSION}/${RHCOS_VERSION}"
+    local i
+    for i in "${RHCOS_IMAGE}" "${RHCOS_KERNEL}" "${RHCOS_INITRAMFS}" ; do
+      if [[ ! -f ${DOWNLOADS_DIR}/${i} ]]; then
+        echo "INFO: download $rhcos_base_url/${i}"
+        wget -nv "$rhcos_base_url/${i}" -O "${DOWNLOADS_DIR}/${i}"
+      fi
+    done
     if [[ ! -f ${DOWNLOADS_DIR}/${LB_IMAGE} ]]; then
-        wget -nv "$LB_IMG_URL" -O "${DOWNLOADS_DIR}/$LB_IMAGE"
+      echo "INFO: download $LB_IMG_URL"
+      wget -nv "$LB_IMG_URL" -O "${DOWNLOADS_DIR}/$LB_IMAGE"
     fi
   fi
 }
@@ -83,4 +88,68 @@ function kubeconfig_copy() {
   mkdir -p ~/.kube
   cp ${KUBECONFIG} ~/.kube/config
   chmod go-rwx ~/.kube/config
+}
+
+function prepare_rhcos_install() {
+  local d="${INSTALL_DIR}/rhcos-install"
+  mkdir -p $d
+  cp "${DOWNLOADS_DIR}/${RHCOS_KERNEL}" "$d/vmlinuz"
+  cp "${DOWNLOADS_DIR}/${RHCOS_INITRAMFS}" "$d/initramfs.img"
+  cat <<EOF > $d/.treeinfo
+[general]
+arch = x86_64
+family = Red Hat CoreOS
+platforms = x86_64
+version = ${OCP_VER}
+[images-x86_64]
+initrd = initramfs.img
+kernel = vmlinuz
+EOF
+}
+
+function prepare_install_config() {
+  local controller_count=$(echo $CONTROLLER_NODES | wc -w)
+  if [ -z "$controller_count" ] ; then
+    echo "ERROR: internal error controller_count must be set"
+    exit 1
+  fi
+  local jinja="$my_dir/../common/jinja2_render.py"
+  $jinja < $my_dir/install_config.yaml.j2 > $INSTALL_DIR/install-config.yaml
+  cat $INSTALL_DIR/install-config.yaml
+}
+
+function patch_ingress_controller() {
+  # By default ingress is scheduled in workers which requires rules in haproxy
+  # that makes more difficult to CI to manage it.
+  # So, patch ingress to re-schedule it on masters
+  local controller_count=$1
+  oc patch ingresscontroller default -n openshift-ingress-operator \
+    --type merge \
+    --patch '{
+      "spec": {
+        "replicas": '${controller_count}',
+        "nodePlacement": {
+          "nodeSelector": {
+            "matchLabels": {
+              "node-role.kubernetes.io/master":""
+            }
+          },
+          "tolerations":[{
+            "effect": "NoSchedule",
+            "operator": "Exists"
+          }]
+        }
+      }
+    }'
+}
+
+function monitor_csr() {
+  local csr
+  while true; do
+    for csr in $(oc get csr 2> /dev/null | grep -w 'Pending' | awk '{print $1}'); do
+      echo "INFO: csr monitor: approve $csr"
+      oc adm certificate approve "$csr" 2> /dev/null || true
+    done
+    sleep 5
+  done
 }
